Move GradingOverview to gradingShape.ts
